+++ vision/src/Color.cpp
@@ -68,37 +68,36 @@ namespace rtx {
     }
   }
 
-  std::pair<unsigned int, unsigned int> getBlobExpectedRealSize(const Color &color) { // TODO: Change to mapping structure?
+  std::pair<double, double> getBlobExpectedRealSize(const Color &color) { // TODO: Change to mapping structure?
     switch (color) {
       case BALL:
-        return std::pair<unsigned int, unsigned int>(BALL_RADIUS * 2, BALL_RADIUS * 2);
+        return std::pair<double, double>(BALL_RADIUS * 2, BALL_RADIUS * 2);
         break;
       case BLUE_GOAL:
       case YELLOW_GOAL:
-        return std::pair<unsigned int, unsigned int>(GOAL_WIDTH, GOAL_HEIGHT);
+        return std::pair<double, double>(GOAL_WIDTH, GOAL_HEIGHT);
         break;
       case FIELD:
-        return std::pair<unsigned int, unsigned int>(10, 10); // TODO: Calibrate
+        return std::pair<double, double>(10, 10); // TODO: Calibrate
         break;
       case ROBOT_YELLOW_BLUE:
       case ROBOT_BLUE_YELLOW:
-        return std::pair<unsigned int, unsigned int>(ROBOT_MAXIMUM_DIAMETER, ROBOT_MARKER_WIDTH * 2); // TODO: Change when detecting arbitrary robot blobs
+        return std::pair<double, double>(ROBOT_MAXIMUM_DIAMETER, ROBOT_MARKER_WIDTH * 2); // TODO: Change when detecting arbitrary robot blobs
         break;
       case WHITE_LINE:
       case BLACK_LINE:
-        return std::pair<unsigned int, unsigned int>(LINE_WIDTH, LINE_WIDTH);
+        return std::pair<double, double>(LINE_WIDTH, LINE_WIDTH);
         break;
       case CHECKERBOARD_WHITE:
       case CHECKERBOARD_BLACK:
       default:
-        return std::pair<unsigned int, unsigned int>(0, 0);
+        return std::pair<double, double>(0, 0);
         break;
     }
   }
 
   std::pair<unsigned int, unsigned int> getBlobExpectedVirtualSize(const Color &color, const std::pair<unsigned int, unsigned int> &position) {
-    std::pair<unsigned int, unsigned int> realSize = getBlobExpectedRealSize(color);
-    // TODO
+    std::pair<double, double> realSize = getBlobExpectedRealSize(color);
     return realSize;
   }
